datasource-hub-mcp/requirements.txt
# MCP Python SDK
mcp>=1.9.2

# Input validation
pydantic>=2.12.3

# HTTP server
uvicorn>=0.30.0
starlette>=0.37.0

# Environment configuration
python-dotenv>=1.0.0

# LLM API (for agent)
# Includes built-in web_search tool support
anthropic>=0.39.0

# HTTP client (for calling external MCP services)
httpx>=0.27.0
